forth

+++ lib/models/shop_app/login_file.dart
@@ -1,4 +1,3 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
 class ShopLoginModel {
   bool? status;
   String? message;
@@ -17,7 +16,7 @@ class UserData {
   String? phone;
   String? image;
   int? points;
-  double? credit;
+  int? credit;
   String? token;
   UserData({
     this.id,
